requirements.txt
python-telegram-bot==20.7
openai
gspread
oauth2client
nest_asyncio